requirements.txt
reachy-sdk==0.5.1
numpy==1.18.5
protobuf==3.20.1
reachy
dynamixel_sdk
pyserial